Fix comparison signed/unsigned mismatch

+++ src/script/nodes/signals/emit_member_signal.cpp
@@ -45,7 +45,7 @@ public:
 
             if (!_method.arguments.empty())
             {
-                if (_args.size() != _method.arguments.size())
+                if (uint64_t(_args.size()) != _method.arguments.size())
                     _args.resize(_method.arguments.size());
 
                 for (int i = 0; i < _method.arguments.size(); i++)
